### 2025-05-27 #8285 @arfon
To JOSS editors, this is a (quick) resubmission from feedback in https://github.com/openjournals/joss-reviews/issues/8229
