### 2025-05-27 #8285 @tiagolbc
@arfon Thank you for your question.

I initially submitted the first version of VibratoScope and the associated paper on April 22, 2025. Since that submission, I have been continuously working on the codebase, performing extensive testing, and implementing a series of improvements—both to the core functionality and to the graphical user interface. I have also been actively refining the paper.md and enhancing the documentation.

Regarding the increase in lines of code from ~1000 to ~2700 in the last 24 hours: This is the result of cumulative work that has been ongoing for several weeks. The recent changes include the addition of new features (such as new pitch extraction methods, expanded batch processing, and advanced plotting/visualization), code refactoring for better modularity, improved error handling, and significant enhancements to the GUI. Additionally, I incorporated feedback from user testing and from my own usage in research, which required the implementation of more comprehensive analysis routines.

The reason these changes appear in a short period is that I was initially under the impression that I should not alter anything in the repository while the submission was under review, unless prompted by the reviewers. Once I realized that it was possible (and even encouraged) to keep improving the project during the review process, I merged all these previously prepared changes into the repository. That’s why there was a rapid growth in code lines and features in the last 24 hours, even though the work itself has been in progress for much longer.

Please let me know if you would have any other questions.

Best regards,

Tiago.
