+++ run.py
@@ -1,5 +1,13 @@
 # run.py
 
+import sys
+from pathlib import Path
+
+# Adicione a pasta src ao sys.path
+current_dir = Path(__file__).parent
+src_dir = current_dir / "src"
+sys.path.insert(0, str(src_dir))
+
 import tkinter as tk
 from gui import VibratoGUI
 from splash import show_splash
