+++ run.py
@@ -3,7 +3,6 @@
 import sys
 from pathlib import Path
 
-# Adicione a pasta src ao sys.path
 current_dir = Path(__file__).parent
 src_dir = current_dir / "src"
 sys.path.insert(0, str(src_dir))
