### 2025-07-09 7bc0172
Delete src/sfeeds.py
