### 2025-07-09 b86e63e
Delete src/core.py
