### 2025-07-09 4eb54fb
Delete src/gui.py
