### 2025-07-09 356df01
Delete src/pitch.py
