+++ src/pitch.py
@@ -1,191 +1,0 @@
-# pitch.py
-
-import os
-import numpy as np
-import pandas as pd
-import librosa
-import parselmouth
-import pyreaper
-from scipy import fft
-from sfeeds import extract_pitch_sfeeds as extract_pitch_with_sfeeds
-
-
-########################################
-# LOAD AUDIO
-########################################
-
-def load_audio(file_path, sr=22050):
-    y, sr = librosa.load(file_path, sr=sr, mono=True)
-    return y, sr
-
-########################################
-# PITCH EXTRACTION FUNCTIONS
-########################################
-
-def extract_pitch_with_praat(wav_path, fmin=50.0, fmax=1500.0, time_step=0.005):
-    """
-    Uses Praat (via Parselmouth) to extract an autocorrelation-based
-    pitch contour, returns (times, freqs), and writes a TXT alongside the WAV.
-    """
-    snd = parselmouth.Sound(wav_path)
-    pitch_obj = snd.to_pitch_ac(
-        time_step=time_step,
-        pitch_floor=fmin,
-        pitch_ceiling=fmax
-    )
-    times = pitch_obj.xs()
-    freqs = pitch_obj.selected_array['frequency']
-    out_txt = os.path.splitext(wav_path)[0] + "_pitch_praat.txt"
-    df = pd.DataFrame({
-        "time_s": times,
-        "f0_Hz": freqs
-    })
-    df.to_csv(out_txt, sep="\t", index=False, header=["time_s", "f0_Hz"])
-    return times, freqs
-
-
-def extract_pitch_with_yin(audio_data, sr, fmin=50.0, fmax=1500.0):
-    """
-    Extract pitch using YIN (librosa.pyin).
-    """
-    f0, _, _ = librosa.pyin(audio_data, sr=sr, fmin=fmin, fmax=fmax)
-    times = librosa.times_like(f0, sr=sr)
-    return times, f0
-
-
-def extract_pitch_with_hps(audio_data, sr, file_path, fmin=50.0, fmax=1500.0, frame_length=0.04, hop_length=0.01):
-    """
-    Extract pitch using Harmonic Product Spectrum (HPS) with harmonic correction.
-    """
-    frame_samples = int(frame_length * sr)
-    hop_samples = int(hop_length * sr)
-    n_frames = int((len(audio_data) - frame_samples) / hop_samples) + 1
-    f0 = np.zeros(n_frames)
-    times = np.arange(n_frames) * hop_length
-
-    for i in range(n_frames):
-        start = i * hop_samples
-        frame = audio_data[start:start + frame_samples]
-        if len(frame) < frame_samples:
-            frame = np.pad(frame, (0, frame_samples - len(frame)), mode='constant')
-
-        window = np.hanning(frame_samples)
-        frame_windowed = frame * window
-
-        N = frame_samples * 4  # Zero-padding
-        spectrum = np.abs(fft.fft(frame_windowed, N))
-        freqs = fft.fftfreq(N, 1 / sr)[:N // 2]
-        spectrum = spectrum[:N // 2]
-
-        hps = spectrum.copy()
-        num_harmonics = 5
-        for h in range(2, num_harmonics + 1):
-            decimated = np.interp(freqs, freqs[::h], spectrum[::h])
-            hps *= decimated
-
-        valid = (freqs >= fmin) & (freqs <= fmax)
-        if not np.any(valid):
-            f0[i] = np.nan
-            continue
-
-        peak_freq = freqs[valid][np.argmax(hps[valid])]
-
-        for h in range(num_harmonics, 1, -1):
-            possible_f0 = peak_freq / h
-            if fmin <= possible_f0 <= fmax:
[…]
