+++ src/core.py
@@ -1,0 +1,163 @@
+# core.py
+
+import numpy as np
+import pandas as pd
+from scipy.signal import find_peaks
+
+########################################
+# VIBRATO DETECTION AND ANALYSIS
+########################################
+
+def detect_vibrato_cycles(times, cents_raw, cents_filtered, prominence=5, distance=5):
+    valid = ~np.isnan(cents_filtered)
+    t_valid = times[valid]
+    filt = cents_filtered[valid]
+    raw = cents_raw[valid]
+
+    peaks, _ = find_peaks(filt, prominence=prominence, distance=distance)
+    troughs, _ = find_peaks(-1 * filt, prominence=prominence, distance=distance)
+    all_idx = np.sort(np.concatenate((peaks, troughs)))
+
+    cycle_times = []
+    cycle_extents = []  # half-extents
+
+    for i in range(len(all_idx) - 1):
+        i1, i2 = all_idx[i], all_idx[i + 1]
+        dt = t_valid[i2] - t_valid[i1]
+        full_amp = abs(filt[i2] - filt[i1])
+        half_amp = full_amp / 2.0
+        cycle_times.append(dt)
+        cycle_extents.append(half_amp)
+
+    return peaks, troughs, np.array(cycle_times), np.array(cycle_extents), t_valid, raw, all_idx
+
+
+def compute_cycle_parameters(times_valid, raw_values, extents, all_idx):
+    cycle_params = []
+    for i in range(len(extents)):
+        idx1 = all_idx[i]
+        idx2 = all_idx[i + 1]
+        t1 = times_valid[idx1]
+        t2 = times_valid[idx2]
+        center_time_s = (t1 + t2) / 2.0  # NOVO!
+        center_cents = (raw_values[idx1] + raw_values[idx2]) / 2.0
+        center_pitch = 2 ** (center_cents / 1200) * 440
+        cycle_params.append({
+            'center_time_s': center_time_s,             # NOVA COLUNA
+            'cycle_time': t2 - t1,
+            'half_extent_cents': extents[i],
+            'center_pitch': center_pitch
+        })
+    return cycle_params
+
+
+def filter_vibrato_cycles(cycle_params, min_half_extent=10.0, max_half_extent=300.0):
+    return [cp for cp in cycle_params if
+            (cp['half_extent_cents'] >= min_half_extent and cp['half_extent_cents'] <= max_half_extent)]
+
+
+def compute_jitter_metrics(periods):
+    N = len(periods)
+    if N < 2:
+        return {
+            'jitter_local_percent': np.nan,
+            'jitter_local_abs_ms': np.nan,
+            'jitter_rap_percent': np.nan,
+            'jitter_ppq5_percent': np.nan,
+            'jitter_ddp_percent': np.nan
+        }
+    mean_period = np.nanmean(periods)
+    diffs = np.abs(np.diff(periods))
+    jitter_local_percent = np.nanmean(diffs) / mean_period * 100
+    jitter_local_abs_ms = np.nanmean(diffs) * 1000
+    if N < 3:
+        jitter_rap_percent = np.nan
+    else:
+        rap_diffs = []
+        for i in range(1, N - 1):
+            avg_three = np.nanmean(periods[i - 1:i + 2])
+            rap_diffs.append(abs(periods[i] - avg_three))
+        jitter_rap_percent = np.nanmean(rap_diffs) / mean_period * 100
+    if N < 5:
+        jitter_ppq5_percent = np.nan
+    else:
+        ppq5_diffs = []
+        for i in range(2, N - 2):
+            avg_five = np.nanmean(periods[i - 2:i + 3])
[…]
