### 2025-07-09 655b9f2
Delete src/gui.py
