+++ src/gui.py
@@ -1,0 +1,1863 @@
+# gui.py
+
+import os
+import sys
+import numpy as np
+import pandas as pd
+import librosa
+import matplotlib
+matplotlib.use('TkAgg')
+import matplotlib.pyplot as plt
+
+import tkinter as tk
+from tkinter import ttk
+from tkinter import filedialog, messagebox, simpledialog
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
+from matplotlib.widgets import SpanSelector
+from pitch import extract_pitch_with_yin, extract_pitch_with_praat, extract_pitch_with_hps, extract_pitch_with_reaper
+
+import sounddevice as sd
+import pyreaper
+import shutil
+
+# Import from your own modules (without dot!)
+from pitch import load_audio, extract_pitch_with_praat, extract_pitch_with_yin, extract_pitch_with_hps, extract_pitch_with_reaper, extract_pitch_with_sfeeds
+from filters import filter_pitch_outliers, apply_bandpass_filter
+from utils import convert_to_cents, remove_mean_or_median, resample_to_uniform_time, sample_entropy
+from spectrogram import plot_spectrogram, plot_before_after_filter, plot_peaks_troughs, final_plot, plot_vibrato_rate
+from core import detect_vibrato_cycles, compute_cycle_parameters, filter_vibrato_cycles, compute_cv, analyze_vibrato, create_vibrato_dataframe, smooth_vibrato_parameters, compute_jitter_metrics
+from audio_recorder import AudioRecorder
+from pathlib import Path
+
+
+########################################
+# MAIN GUI CLASS
+########################################
+class VibratoGUI:
+    def __init__(self, root):
+        from PIL import Image, ImageTk
+        import tkinter as tk
+        from tkinter import font as tkFont
+        from tkinter import ttk  # Import ttk from tkinter
+        from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
+        import matplotlib.pyplot as plt
+
+        self.root = root
+        self.root.title("Vibrato Scope - version 1.1.0")
+        self.root.configure(bg="#b0afa6")
+
+        # Fonts
+        cal_font = tkFont.Font(family="Calibri", size=11)
+        cal_bold = tkFont.Font(family="Calibri", size=12, weight="bold")
+
+        # Control variables
+        self.audio_data = None
+        self.recording_stopped = False
+        self.sr = None
+        self.times = None
+        self.pitch_hz = None
+        self.t_uniform = None
+        self.cents_uniform = None
+        self.selected_regions = []
+        self.file_path = ""
+        self.df_detailed = None
+        self.df_avg = None
+        self.df_region = None
+        self.summary_data = None
+        self.last_fig = None
+        self.generated_figures = []
+        self.alternative_used = False
+        self.harmonic_used = None
+        self.harmonic_frequency = None
+        self.spectrogram_type = tk.StringVar(value="standard")
+        self.pitch_method = tk.StringVar(value="praat")
+        self.actual_pitch_method = tk.StringVar(value="praat")
+        self.sampen_target = tk.StringVar(value="rate")
+        self.pitch_method.trace_add("write", self.on_pitch_method_change)
+        self.sampen_distance = tk.StringVar(value="chebyshev")
+
+        # --- TOPO: Logo + controls ---
+        self.top = tk.Frame(root, bg="#b0afa6")
+        self.top.grid(row=0, column=0, sticky="ew", padx=10, pady=2)
+
+        try:
+            logo_image = Image.open("logo.png")
[…]
