+++ docs/paper.md
@@ -79,7 +79,7 @@ Figure \autoref{fig:test_example} illustrates the analysis of a synthetic vowel 
 
 # Statement of Need
 
-Vibrato is a cornerstone of vocal expression, yet its quantitative analysis is often constrained by the limitations of existing tools. Proprietary software like VoceVista Pro provides real-time vibrato overlays on spectrograms but is hindered by opaque algorithms, limited export options, and commercial licensing costs, reducing transparency and reproducibility [@vocevista:2022]. BioVoice offers high-resolution estimates of vibrato rate, extent, jitter, and shimmer, but its Windows-only executable lacks batch processing and an API, limiting integration with automated pipelines [@morelli:2019]. Open-source alternatives, such as the Embodied Music Lab (EML) Praat script, are cross-platform and freely available but require manual configuration and execution within Praat’s scripting console, posing barriers for users without coding expertise [@howell:2024]. The Vibrato Analysis Toolbox (VAT) provides a sophisticated Hilbert-transform pipeline with user-definable filters, yet its reliance on MATLAB ties it to costly licenses and demands signal-processing expertise, restricting its accessibility [@zhang:2017].
+Vibrato is a cornerstone of vocal expression, yet its quantitative analysis is often constrained by the limitations of existing tools. Proprietary software like VoceVista Pro provides real-time vibrato overlays on spectrograms but is hindered by opaque algorithms, limited export options, and commercial licensing costs, reducing transparency and reproducibility [@vocevista:2022]. BioVoice offers high-resolution estimates of vibrato rate, extent, jitter, and shimmer, but its Windows-only executable lacks batch processing and an API, limiting integration with automated pipelines [@morelli:2019]. Open-source alternatives, such as the Embodied Music Lab (EML) Vibrato Tools [@howell:2025], are cross-platform Praat plugins that provide a user-friendly graphical interface and do not require coding skills or manual scripting. These tools offer default configurations that work for most cases. While the need to install Praat and the plugin may represent an additional step, the EML tools are highly accessible and well-documented. The Vibrato Analysis Toolbox (VAT) provides a sophisticated Hilbert-transform pipeline with user-definable filters, yet its reliance on MATLAB ties it to costly licenses and demands signal-processing expertise, restricting its accessibility [@zhang:2017].
 
 VibratoScope was designed to address these gaps, offering a comprehensive, open-source solution for high-throughput vibrato analysis. Implemented in Python and released under an MIT license, it combines: (i) an intuitive graphical user interface (GUI) for interactive audio selection, pitch extraction, and vibrato cycle detection, accessible to non-programmers; (ii) a command-line interface supporting unattended batch processing with multiple pitch extraction methods, including Praat [@boersma:1993], YIN [@decheveigne:2002], Harmonic Product Spectrum [@noll:1970], and REAPER [@talkin:2015]; and (iii) transparent CSV and PNG outputs that integrate seamlessly with statistical and machine-learning workflows. By merging the ease-of-use of commercial GUI software, the extensibility of script-based tools, and the accessibility of open-source code, VibratoScope enables reproducible research for voice scientists, pedagogues, and clinicians. Its class-based API allows developers to extend functionality, while its platform-independent design supports diverse applications, from vocal pedagogy to cross-cultural studies of singing styles. Adopted in academic settings, including graduate courses and workshops at conferences like the Pan-European Voice Conference, VibratoScope meets the growing demand for open, scalable tools in voice research.
 
