### 2025-07-09 1e25ee4
Update paper.bib
