+++ docs/paper.bib
@@ -44,11 +44,12 @@
   howpublished = {"https://www.vocevista.com/"}
 }
 
-@misc{howell2024,
-  author    = {Howell, Ian},
-  title     = {EML Vibrato Tools},
-  year      = {2024},
-  howpublished = {"https://www.embodiedmusiclab.com/member-eml-vibrato-tools"}
+@misc{howell2025,
+  author    = {Ian Howell and Tzvetana Nestorova},
+  title     = {{EML Vibrato Tools}},
+  year      = {2025},
+  note      = {Published online January 22, 2025},
+  howpublished = {\url{https://www.embodiedmusiclab.com/eml-tools-sign-up}}
 }
 
 @article{zhang2017,
