### 2025-07-09 05d0f2a
Update paper.bib
