+++ docs/paper.bib
@@ -45,7 +45,7 @@
 }
 
 @misc{howell2025,
-  author    = {Ian Howell and Tzvetana Nestorova},
+  author    = {Ian Howell and Theodora Nestorova},
   title     = {{EML Vibrato Tools}},
   year      = {2025},
   note      = {Published online January 22, 2025},
