### 2025-10-30 #8285 @samhforbes
@tiagolbc just looking at this I think the references might be the issue? I can see they're being referenced as eg @seashore:1932, but the bib file has it as eg seashore1932. Do you want to try starting there?
