### 2025-10-31 3e26c40
Update paper.md
