### 2025-10-31 5dd270b
Update paper.md
