### 2025-10-31 3e02f91
Update paper.md
