### 2025-10-31 e1540ac
Update paper.md
