### 2026-02-19 #10085 @editorialbot
[REVIEW]: VibratoScope: A Python Toolkit for High-Resolution Vibrato Analysis in Singing Voice
**Submitting author:** @tiagolbc (<a href="http://orcid.org/0000-0002-8355-5436">Tiago Cruz</a>)
**Repository:** https://github.com/tiagolbc/vibratoscope
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
