### 2026-03-09 tiagolbc/vibratoscope#5 @ax-le
REVIEW for JOSS submission.
You may find here my review for your submission to JOSS.

First of all, this toolbox is an interesting and valuable tool for vocal analysis. As mentioned in your Statement of Need, providing an easy-to-use and efficient vibrato analysis toolbox is a valuable contribution, in particular when compared with Windows-only or MATLAB alternatives.

During my review, I found a few areas in both the software repository and the paper that need addressing to meet JOSS guidelines and improve usability.

[EDIT]: This review was updated considering the first pass of modifications provided by the main author.
[EDIT2]: This review was updated a 2nd time, considering the second pass of modifications provided by the main author.

- Software and Repository Issues:
   - [x] ~~Installation Dependencies: The installation does not work without additional package installations not currently covered (Note: This has already been mentioned in other issues #1 and #2). It may even be very complex to install it naively because `pkg_resources` was removed from the most recent version of `setuptools` (since v.81), leading to the standard `pip install setuptools` installation not working.~~ Done.
   - [x] ~~Audio Library Alternative: I don't understand the need to use PortAudio, which requires an additional manual installation. In my opinion, it could be replaced with soundfile and sounddevice with no loss of functionality.~~ The author responded and decided not to change it. While I find this solution a little disappointing, I do understand. After all, this is an upgrade, not a deal-breaker.
   - [x] ~~CLI Documentation: There is no documentation detailing how to use the Command Line Interface (CLI).~~ ~~It appears that mentions to the CLI was removed in the new version, I guess that CLI is not present anymore?~~ The author removed all references to API/CLI integration, clarifying that this is not present in the current submission. This point does not appear valid anymore.
   - [x] ~~API and main Integration: ~~While the GUI is functional, there is no native Python main function or an integrated API that makes it easy to use VibratoScope programmatically within Python scripts. While I acknowledge that this is not mentioned in the paper, I believe that this is a real problem in the easy integration of the toolbox into automated packages and scripts. Notably, it substantially weakens the statement "VibratoScope supports both interactive and automated workflows" (l.32).~~ The author removed the part about "automated workflows, which suggests that it is not possible to use the toolbox without the GUI. I think that it should be mentioned early in the paper, maybe even in the first lines of the README.~~ The author removed all references to API/CLI integration, clarifying that this is not present in the current submission. This point does not appear valid anymore.
   - [x] ~~Missing Tests: There do not appear to be any automated tests in the repository, which is a requirement in the JOSS checklist.~~ ~~Tests are now present, but naively running them leads to errors ("ModuleNotFoundError: No module named 'core'" for `test_core.py`, and ModuleNotFoundError: No module named 'filters' for `test_filters_utils.py`). An additional installation step (like installing the package and referencing the module name in the tests) seems necessary.~~ All tests pass.
   - [x] ~~Lack of DNN-based methods: Are there plans to include recent DNN-based f0​-estimation methods? It would be worth mentioning why they are omitted or if they are slated for future releases, as they represent the State-of-the-Art models.~~ The author clarified that this point is out-of-scope.
   - [x] ~~SFEEDS Reference: What is SFEEDS? It is mentioned in the README but has no reference or explanation, and it is not mentioned in the paper at all.~~ Done.
   - [x] ~~Figure Naming: When running the code and saving results, some figures are just stored/captioned loosely as "figure 1, figure 2", etc.~~ Done.

- Paper Structure and Content
   - [x] ~~Statement of Need Placement: I recommend moving the "Statement of Need" section earlier in the paper. In JOSS papers, this is typically placed right after the Summary to immediately justify the software's existence before diving into technical details.~~ Done.
   - [x] ~~Unjustified Claim: The sentence "These test cases confirm that VibratoScope reliably measures vibrato parameters across known input conditions, validating its use for both research and pedagogical applications" (l. 53/54) feels overstated and is not rigorously justified by the preceding text because it only presents results of the toolboxes on individual and simple (voice-only, low noise) examples. Consider softening this claim or providing more concrete backing.~~ Done (removed).
   - [x] ~~Related Work Relevance: The State-of-the-Art and related work need to be more explicit (e.g., a dedicated Related Work section) and better introduced. For instance, the transition to "These findings reinforce the view of vibrato as a stylistically dependent and context-sensitive vocal parameter" is interesting, but it does not clearly connect to why this software tool is technically necessary. The motivations for the choices of the f_0 estimation algorithms are not provided.~~ Done: better presentation of the related work.
   - [x] ~~Comparison to EML: The paper mentions the Embodied Music Lab (EML) open-source toolbox but fails to explicitly compare VibratoScope against it. In particular, the sole criticism of EML (that it requires installing Praat) feels insufficient to justify the need for a completely new tool.~~ Done. It's the same justification, but it's more convincing as it is formulated now imo.
   - [x] ~~Undefined Concepts: Several domain-specific concepts are introduced without definition (e.g., sample entropy, determinism, and line length) , or are not detailed enough for a general scientific audience (e.g., Jitter and Shimmer).~~
   - [x] ~~The "Mathematics" Section: The section title is slightly inelegant, and the section in itself lacks details. For instance, stating that vibrato rate is "typically computed via Fourier analysis or autocorrelation" is a bit vague for a section dedicated to mathematical implementation. It should present all metrics used in the toolbox.~~ Done (removed).

- Typos and Formatting
   - [x] ~~Broken Reference: On line 55, there is a broken figure reference: "Figure ?? illustrates the analysis...".~~ Done, reference deleted in the new version.
   - [x] ~~Math Formatting: There are Markdown/Latex formatting issues on lines 89 and 91, currently written as (f_0(t)) and (fmean​) instead of standard inline math formatting.~~ Done.
   - [x] ~~Outdated Script Name: The paper references `vibratoscope.py` for installation/usage, which appears to no longer be true based on the current repository structure.~~ Done.

I look forward to seeing how VibratoScope evolves; with these adjustments, it will undoubtedly be a fantastic asset to the vocal analysis community!

----
## EDIT: 2nd round of review.

The paper was improved between the two rounds of review. Congrats!

The main criticism I had was the fact that the toolbox could not be installed easily. This was solved, and the toolbox may now be installed with the specified requirements.

There is still a problem with the tests that cannot be run (at least on my machine). Please see the dedicated item above.

I think that the tollbox could be improved with the items that remain open, but they generally concern additional functionalities, or modificaions, not mentioned by the author. Hence, this may be seen as future work if the authors find these points relevant.

- New points raised in the 2nd round of reviews:
   - [x] ~~Falling installation: Installation may fail due to pyreaper on Windows, due to an additional dependency (Microsoft Tools for C++ app). Because this dependency is optional, I suggest removing it from the requirements.txt and adding it only as a comment.~~ Done, `pyreaper` is now an optional dependency.

----
## EDIT: 3rd round of review.

All the points raised in my previous rounds of review were tackled by the author. I am thus closing my review.
