### 2026-03-10 tiagolbc/vibratoscope#5 @hadware
I totally second most (if not all) of these review points. Thanks a lot @ax-le for this extremely detailed review
