### 2026-04-17 474abba
Refine vibrato analysis descriptions and structure
Updated the paper to enhance clarity and precision in the description of vibrato analysis, including corrections in terminology and improved structure for better readability.
