### 2026-04-17 38f0fe9
Update paper.md
