+++ docs/paper.md
@@ -98,13 +98,13 @@ where $m$ is the pattern length, $r$ is the tolerance, $N$ is the number of data
 
 # Figures
 
-![Graphical User Interface](../figures/gui.png){#fig:gui width=90%}
+![Graphical User Interface](../figures/gui.png)
 
-![Pitch Filtering](../figures/pitch_filtering.png){#fig:pitch_filtering width=90%}
+![Pitch Filtering](../figures/pitch_filtering.png)
 
-![Peak and Trough Detection](../figures/peak_trough_detection.png){#fig:peak_trough_detection width=90%}
+![Peak and Trough Detection](../figures/peak_trough_detection.png)
 
-![Final Analysis Summary](../figures/Figure_Analysis_Vibrato.png){#fig:final_analysis width=90%}
+![Final Analysis Summary](../figures/Figure_Analysis_Vibrato.png)
 
 # Acknowledgements
 
