### 2026-04-17 a94f4c0
Add new article reference for frequency-detection algorithm
