+++ docs/paper.bib
@@ -164,6 +164,15 @@
   doi       = {10.1121/1.1911939}
 }
 
+@article{kitayama2024,
+  author    = {Itsuki Kitayama and Kiyohito Hosokawa and Shinobu Iwaki and Misao Yoshida and Akira Miyauchi and Toshihiro Kishikawa and Hidenori Tanaka and others},
+  title     = {Robust fundamental frequency-detection algorithm unaffected by the presence of hoarseness in human voice},
+  journal   = {The Journal of the Acoustical Society of America},
+  volume    = {156},
+  number    = {6},
+  pages     = {4217--4228},
+  year      = {2024}
+}
 @misc{talkin2015,
   author    = {Talkin, David},
   title     = {REAPER: Robust epoch and pitch estimator},
