### 2026-04-17 976b385
Remove duplicate REAPER citation from paper.bib
