+++ docs/paper.bib
@@ -164,6 +164,13 @@
   doi       = {10.1121/1.1911939}
 }
 
+@misc{talkin2015,
+  author    = {Talkin, David},
+  title     = {REAPER: Robust epoch and pitch estimator},
+  year      = {2015},
+  howpublished = {\url{https://github.com/google/REAPER}}
+}
+
 @article{kitayama2024,
   author    = {Itsuki Kitayama and Kiyohito Hosokawa and Shinobu Iwaki and Misao Yoshida and Akira Miyauchi and Toshihiro Kishikawa and Hidenori Tanaka and others},
   title     = {Robust fundamental frequency-detection algorithm unaffected by the presence of hoarseness in human voice},
@@ -172,10 +179,4 @@
   number    = {6},
   pages     = {4217--4228},
   year      = {2024}
-}
-@misc{talkin2015,
-  author    = {Talkin, David},
-  title     = {REAPER: Robust epoch and pitch estimator},
-  year      = {2015},
-  howpublished = {\url{https://github.com/google/REAPER}}
 }
