### 2026-04-17 88ec4d4
Update audio_recorder.py
