+++ src/audio_recorder.py
@@ -1,9 +1,5 @@
-# audio_recorder.py
-
 import tkinter as tk
 from tkinter import ttk, filedialog, messagebox
-import pyaudio
-import wave
 import threading
 import numpy as np
 import time
@@ -14,6 +10,9 @@ import psutil
 import shutil
 import os
 import queue
+import sounddevice as sd
+import soundfile as sf
+
 
 class AudioRecorder(tk.Toplevel):
     def __init__(self, parent, on_save=None):
@@ -26,17 +25,17 @@ class AudioRecorder(tk.Toplevel):
         self.tmp_file = "temp_recording.wav"
         self.sr = None
         self.duration = 5
-        self.chunk = 2048  # Increased for stability
+        self.chunk = 2048
         self.rec_thread = None
         self.rec_exit_event = None
         self.live_plot = False
         self.device_index = None
         self.plot_mode = "Raw"
-        self.db_range = 80  # Default to ±80 dB for voice
+        self.db_range = 80
         self.audio_queue = queue.Queue()
         self.plot_data = []
+        self.recorded_chunks = []
 
-        # Setup logging
         logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
         self.logger = logging.getLogger(__name__)
 
@@ -50,13 +49,20 @@ class AudioRecorder(tk.Toplevel):
         cal_font = tkFont.Font(family="Calibri", size=11) if "Calibri" in tkFont.families() else tkFont.Font(
             family="Arial", size=11)
 
-        # Header
-        tk.Label(self, text="Before recording select the options below.",
-                 bg="#b0afa6", font=(cal_font.actual("family"), 14, "bold")).pack(pady=(20, 5))
+        tk.Label(
+            self,
+            text="Before recording select the options below.",
+            bg="#b0afa6",
+            font=(cal_font.actual("family"), 14, "bold")
+        ).pack(pady=(20, 5))
 
-        tk.Label(self, text="Don't forget to increase Max Duration if you want to sing a song", bg="#b0afa6", font=(cal_font.actual("family"), 12, "italic")).pack()
+        tk.Label(
+            self,
+            text="Don't forget to increase Max Duration if you want to sing a song",
+            bg="#b0afa6",
+            font=(cal_font.actual("family"), 12, "italic")
+        ).pack()
 
-        # Input device
         device_frame = tk.Frame(self, bg="#b0afa6")
         device_frame.pack(pady=5)
         tk.Label(device_frame, text="Input Device:", bg="#b0afa6", font=cal_font).pack(side=tk.LEFT, padx=5)
@@ -64,7 +70,6 @@ class AudioRecorder(tk.Toplevel):
         self.device_menu = ttk.Combobox(device_frame, textvariable=self.device_var, state="readonly", width=50)
         self.device_menu.pack(side=tk.LEFT)
 
-        # Sample Rate
         sr_frame = tk.Frame(self, bg="#b0afa6")
         sr_frame.pack(pady=2)
         tk.Label(sr_frame, text="Sample Rate (Hz):", bg="#b0afa6", font=cal_font).pack(side=tk.LEFT, padx=5)
@@ -72,7 +77,6 @@ class AudioRecorder(tk.Toplevel):
         self.sr_menu = ttk.Combobox(sr_frame, textvariable=self.sr_var, state="readonly", width=10)
         self.sr_menu.pack(side=tk.LEFT)
 
-        # Buffer Size
         buf_frame = tk.Frame(self, bg="#b0afa6")
         buf_frame.pack(pady=2)
[…]
