### 2026-04-17 6a648fa
Update spectrogram.py
