### 2026-04-17 81aa52c
Remove duplicate installation notes in README
Removed duplicate lines about pyaudio and REAPER support.
