### 2026-04-17 #10085 @tiagolbc
> [@tiagolbc](https://github.com/tiagolbc) do you have some updates?

Dear @faroit ,

Thank you very much for the careful and constructive review. I have revised the repository and the paper accordingly.

The main changes are as follows:

I revised the paper structure and wording, including moving the Statement of Need earlier, clarifying the scope of the software, and removing wording that could suggest a command-line workflow when the current release is centered on GUI-based analysis and GUI-based batch processing.
I expanded the description of the implemented metrics and clarified definitions of jitter, sample entropy, determinism, and line length for a broader scientific audience.
I corrected the mathematical formatting issues in the manuscript and revised the implementation-focused section to better describe how the metrics are computed.
I updated the discussion of related tools, including a more accurate description of the EML Vibrato Tools and a clearer explanation of how VibratoScope differs, especially in providing a standalone GUI environment with batch processing across multiple files.
I added automated tests for the scientific core of the software. The repository now includes a tests/ directory with pytest-based tests covering core analysis functions.
I improved the repository documentation, including installation instructions and clarification of optional dependencies.
I revised figure naming so that generated outputs now use descriptive filenames rather than generic names such as “Figure 1” or “Figure 2.”
I updated the paper and README to reflect the current execution entry point and repository organization.

Thank you again for the detailed feedback. It helped substantially improve both the manuscript and the repository.
