### 2026-04-17 #10085 @faroit
@hadware given the changes, could you continue with your review now?
