### 2026-05-05 tiagolbc/vibratoscope#5 @faroit
@ax-le can you please check which of the items from your review have been addressed so far and update us on the main thread? @tiagolbc
