### 2026-05-23 tiagolbc/vibratoscope#5 @ax-le
Hi @faroit !
I have just updated my first review, in the message directly. I still have some concerns, but most have been addressed. Overall, I think the toolbox and the paper are aligned, and I find the toolbox easy to use.
I have also updated my checklist in the main thread; only the tests and some info on how to contribute seem to be missing, in my opinion.
Let me know if you have questions or comments!
Best,
ax-le.
