### 2026-06-12 c4e695e
Delete tests/test_core.py
