+++ requirements.txt
@@ -6,7 +6,6 @@ pandas
 sounddevice
 soundfile
 praat-parselmouth
-pyreaper
 Pillow
 psutil
 pyaudio
