### 2026-06-12 #10085 @tiagolbc
Dear @ax-le, @hadware @faroit

Thank you again for the careful second-round review and for updating the checklist.

I have now addressed the remaining blocking issue concerning the automated tests. The tests were already present, but the import structure was not robust enough when running them in a clean environment. I updated the test imports to use the package-style structure, for example `from src.core import ...`, added `src/__init__.py`, and added a root-level `pytest.ini`.

The tests should now be run from the repository root with:

```bash
python -m pytest -q
```

This should resolve the previous `ModuleNotFoundError` issues related to `core`, `filters`, and other modules.

Regarding the installation issue raised in the second round, I agree that `pyreaper` may create installation difficulties on Windows because it can require Microsoft C++ Build Tools. To avoid blocking the basic source-code installation, I have removed `pyreaper` from the main `requirements.txt`.

However, REAPER support remains available in VibratoScope as an optional pitch-extraction method. The code already treats `pyreaper` as optional: if it is not installed, the program only raises an error when the user explicitly selects the REAPER method. This allows the standard installation and the automated tests to run without requiring `pyreaper`.

For the packaged executable version of VibratoScope, `pyreaper` can still be bundled during the build process. Therefore, users who download the executable can continue to access the REAPER method without needing to install it manually.

Regarding the audio library suggestion, I have decided not to change the audio backend during the current JOSS review cycle. I understand the suggestion to avoid PortAudio-related dependencies, but changing the audio I/O stack at this stage could introduce new compatibility issues without directly affecting the core vibrato-analysis functionality described in the paper. The main analytical workflow is based on loading audio files, extracting pitch contours, filtering vibrato modulation, detecting cycles, and exporting CSV/PNG outputs. For this reason, I prefer to keep the current implementation stable for this release and consider backend simplification in a future version.

Regarding DNN-based pitch-estimation methods, I do not plan to include them in the current release. This is an intentional scope decision. VibratoScope currently prioritizes established and interpretable pitch-extraction methods commonly used in voice and singing research. DNN-based methods are certainly relevant, but they would require additional evaluation, validation, and documentation before being responsibly integrated into this toolbox. I therefore consider this future work rather than a requirement for the present release.

Regarding CLI/API support, I revised the wording to avoid implying that VibratoScope currently provides a stable command-line interface or public Python API. The current version is centered on GUI-based interactive analysis and GUI-based batch processing. I agree that a programmatic API could be useful in the future, but it is outside the scope of the current JOSS submission.

The other points raised in the first review were addressed in the previous revision, including:

* moving the Statement of Need earlier in the paper;
* improving the presentation of related work;
* clarifying the comparison with existing tools;
* removing or softening unsupported validation claims;
* correcting the outdated script name;
* fixing broken figure references and math formatting;
* clarifying SFEEDS;
* improving figure naming;
* clarifying that the current release does not include DNN-based pitch estimators.

Regarding contribution information, I acknowledge this point. For now, issues and pull requests can be submitted directly through the GitHub repository, and I will keep improving the repository documentation as the project evolves.

Thank you again for the detailed and constructive review. I hope the latest changes resolve the remaining blocking issue related to the automated tests while keeping the scope of the current release clear.

Best,
Tiago
