### 2026-06-16 tiagolbc/vibratoscope#5 @ax-le
Hi @faroit !
I've just finished my second review, taking into account the author's response. I think the author has addressed all of my points, so I am closing my review.
Best,
ax-le.
