### 2026-07-08 e6b163e
Add community guidelines to README
Added community guidelines for contributions, issue reporting, and support requests.
