+++ README.md
@@ -81,6 +81,65 @@ These examples are used in validation and reproducibility. See `docs/paper.md` f
 
 ---
 
+## 🤝 Community Guidelines
+
+VibratoScope welcomes contributions, bug reports, feature requests, and questions from the community.
+
+### Contributing to the software
+
+Contributions are welcome through GitHub pull requests.
+
+Before starting a substantial change, please open an issue to describe the proposed modification and discuss whether it fits the scope of the project. This helps avoid duplicated work and ensures that new features remain consistent with VibratoScope’s goals.
+
+To contribute:
+
+1. Fork the repository.
+2. Create a new branch for your change.
+3. Make your changes with clear and readable code.
+4. Include documentation updates when relevant.
+5. If your change affects the analysis workflow, include a minimal example or describe how the change can be tested.
+6. Submit a pull request explaining the motivation and the main changes.
+
+Contributions may include bug fixes, documentation improvements, validation examples, new analysis features, improvements to the GUI, or refinements to existing pitch-tracking and vibrato-analysis routines.
+
+### Reporting issues or problems
+
+Bugs, unexpected results, installation problems, or documentation errors should be reported using GitHub Issues.
+
+When opening an issue, please include as much of the following information as possible:
+
+- Operating system
+- Python version
+- VibratoScope version or commit hash
+- Installation method
+- Pitch extraction method used, if relevant
+- A clear description of the problem
+- Steps needed to reproduce the issue
+- Expected behavior
+- Actual behavior
+- Error messages or screenshots, when available
+- A small example file, if possible and ethically shareable
+
+Please do not upload audio files containing identifiable or sensitive voice data unless you have permission to share them publicly.
+
+### Seeking support
+
+For questions about installation, use, interpretation of outputs, or reproducibility, please open a GitHub Issue using a clear title.
+
+Support requests may include questions about:
+
+- Installing dependencies
+- Running the GUI
+- Batch processing
+- Pitch extraction methods
+- Interpreting CSV outputs
+- Reproducing the example analyses
+- Using VibratoScope in research workflows
+
+For academic use, please also consult the documentation and examples provided in this repository before opening a support request.
+
+---
+
 ## 📖 Citation
 
 If you use this toolkit in research, please cite:
