+++ src/core.py
@@ -5,49 +5,264 @@ import pandas as pd
 from scipy.signal import find_peaks
 
 ########################################
+# INTERNAL HELPERS
+########################################
+
+class _ExtremaIndexArray(np.ndarray):
+    """
+    Integer extrema indices carrying refined sub-sample metadata.
+
+    The class behaves as a normal NumPy integer array, so the legacy GUI and
+    existing callers remain compatible. The additional metadata allows
+    compute_cycle_parameters() to use the same refined extrema times and raw
+    contour values already calculated by detect_vibrato_cycles().
+    """
+
+    def __new__(cls, values, refined_times=None, refined_raw_values=None):
+        obj = np.asarray(values, dtype=int).view(cls)
+        obj.refined_times = (
+            None if refined_times is None
+            else np.asarray(refined_times, dtype=float)
+        )
+        obj.refined_raw_values = (
+            None if refined_raw_values is None
+            else np.asarray(refined_raw_values, dtype=float)
+        )
+        return obj
+
+    def __array_finalize__(self, obj):
+        if obj is None:
+            return
+        self.refined_times = getattr(obj, "refined_times", None)
+        self.refined_raw_values = getattr(obj, "refined_raw_values", None)
+
+
+def _quadratic_extremum_time(times, values, index):
+    """
+    Refine an extremum time by fitting a parabola to the point and its
+    immediate neighbours.
+
+    The calculation uses the actual time coordinates, so it remains valid even
+    when the nominal 100 Hz grid is not represented by exactly 10 ms steps.
+    """
+    times = np.asarray(times, dtype=float)
+    values = np.asarray(values, dtype=float)
+    index = int(index)
+
+    if index <= 0 or index >= len(values) - 1:
+        return float(times[index])
+
+    x = times[index - 1:index + 2]
+    y = values[index - 1:index + 2]
+
+    if not (np.all(np.isfinite(x)) and np.all(np.isfinite(y))):
+        return float(times[index])
+
+    # Centre x values to improve numerical conditioning.
+    x0 = float(x[1])
+    xc = x - x0
+
+    try:
+        a, b, _ = np.polyfit(xc, y, 2)
+    except Exception:
+        return float(times[index])
+
+    if not np.isfinite(a) or not np.isfinite(b) or abs(a) < 1e-15:
+        return float(times[index])
+
+    vertex_offset = -b / (2.0 * a)
+    lower = float(xc[0])
+    upper = float(xc[-1])
+
+    if not np.isfinite(vertex_offset) or vertex_offset < lower or vertex_offset > upper:
+        return float(times[index])
+
+    return x0 + float(vertex_offset)
+
+
+def _quadratic_value_at_time(times, values, index, target_time):
+    """
+    Evaluate the local raw contour at a refined extremum time.
+
+    A three-point quadratic interpolation is preferred. Linear interpolation
+    over the finite contour is used as a safe fallback.
+    """
+    times = np.asarray(times, dtype=float)
+    values = np.asarray(values, dtype=float)
+    index = int(index)
+    target_time = float(target_time)
+
+    if 0 < index < len(values) - 1:
+        x = times[index - 1:index + 2]
+        y = values[index - 1:index + 2]
+
+        if np.all(np.isfinite(x)) and np.all(np.isfinite(y)):
+            x0 = float(x[1])
+            xc = x - x0
+            try:
+                coeffs = np.polyfit(xc, y, 2)
+                value = np.polyval(coeffs, target_time - x0)
[…]
