### 2026-07-20 17d6f82
Update gui.py
