### 2026-07-20 a6d823c
Update utils.py
