+++ src/utils.py
@@ -25,37 +25,47 @@ def frequency_to_note_name(freq, ref_hz=440.0):
 
 def sample_entropy(time_series, m=2, r=None, distance='chebyshev'):
     """
-    Computes the Sample Entropy (SampEn) of a time series.
+    Computes the conventional Sample Entropy (SampEn) of a time series.
 
     Parameters:
         time_series (array-like): Input time series data.
         m (int, optional): Length of compared sequences (default: 2).
-        r (float, optional): Tolerance for distance (default: 0.2 * std of series).
-        distance (str, optional): Distance metric ('chebyshev', 'euclidean', 'manhattan'; default: 'chebyshev').
+        r (float, optional): Tolerance for distance
+            (default: 0.2 * standard deviation of the series).
+        distance (str, optional): Distance metric
+            ('chebyshev', 'euclidean', 'manhattan'; default: 'chebyshev').
 
     Returns:
-        float: SampEn value, np.nan if insufficient data, np.inf if no matches.
+        float: SampEn value, np.nan if data are insufficient or non-finite,
+        and np.inf if no valid matches exist.
 
     Notes:
-        - SampEn = -ln(A(m+1) / A(m)), where A(m) is the proportion of vector pairs of length m within r.
-        - Data is not normalized internally; provide normalized input if scale invariance is needed.
-        - Bias correction is applied for small sample sizes using an approximate factor (m+1)/(N-m).
-        - Requires at least m+1 points; returns np.nan if unmet.
+        - SampEn = -ln(A(m+1) / A(m)).
+        - Self-matches are excluded.
+        - No additional multiplicative bias factor is applied.
+        - Data are not normalized internally; provide normalized input when
+          scale invariance is required.
     """
-    time_series = np.asarray(time_series, dtype=np.float64)
+    time_series = np.asarray(time_series, dtype=np.float64).ravel()
     N = len(time_series)
 
     if m < 1:
         raise ValueError("m must be at least 1")
-    if N < m + 1:
-        return np.nan  # Not enough data
+    if N < m + 2:
+        return np.nan
+    if not np.all(np.isfinite(time_series)):
+        return np.nan
 
     if r is None:
-        r = 0.2 * np.std(time_series)
-    if r <= 0:
-        raise ValueError("r must be positive")
+        series_std = np.std(time_series)
+        if series_std == 0:
+            return 0.0
+        r = 0.2 * series_std
 
-        # Define distance function
+    r = float(r)
+    if not np.isfinite(r) or r <= 0:
+        raise ValueError("r must be a positive finite number")
+
     if distance == 'chebyshev':
         dist_func = lambda x, y: np.max(np.abs(x - y), axis=1)
     elif distance == 'euclidean':
@@ -66,27 +76,25 @@ def sample_entropy(time_series, m=2, r=None, distance='chebyshev'):
         raise ValueError(f"Unknown distance metric: {distance}")
 
     def _phi(m_len):
-        patterns = np.array([time_series[i:i + m_len] for i in range(N - m_len + 1)])
+        patterns = np.array(
+            [time_series[i:i + m_len] for i in range(N - m_len + 1)],
+            dtype=np.float64
+        )
         count = 0
         for i in range(len(patterns)):
             dists = dist_func(patterns, patterns[i])
-            count += np.sum(dists <= r) - 1  # exclude self-match
-        denom = (N - m_len + 1) * (N - m_len)
-        return count / denom if denom > 0 else 0
[…]
