### 2026-07-21 a7a8d3b
Update gui.py
